Réduit le scope du plugin Cantera à PFR uniquement
- Supprime equilibrium_composition, reactor_profile, flame_speed,
  thermo_properties, kinetic_sensitivity, spectral_coupling,
  read_cantera_mechanism du plugin (préservés dans _legacy.py)
- Enregistre PFR comme contribution simulation dans le namespace cantera
- Passe le stub core PFR de ImportError à MissingPluginError
  (message clair : pip install spectrochempy-cantera)
- Nettoie les hooks : register_analyses et register_readers vides
- Ajuste les tests plugin à la new scope PFR-only
- 198 core tests + 9 plugin tests pass, ruff clean
