+++ plugins/spectrochempy-cantera/tests/test_cantera.py
@@ -1,31 +1,22 @@
 # ruff: noqa: S101  # assert allowed in tests
 
-"""Tests for spectrochempy-cantera plugin."""
+"""Tests for spectrochempy-cantera plugin — PFR scope only."""
 
 from __future__ import annotations
-
-from pathlib import Path
 
 import pytest
 from spectrochempy_cantera import PFR
 from spectrochempy_cantera import CanteraPlugin
-from spectrochempy_cantera import equilibrium_composition
-from spectrochempy_cantera import reactor_profile
-from spectrochempy_cantera import thermo_properties
 
 from spectrochempy.api.plugins import PluginCapability
 from spectrochempy.api.plugins import PluginState
 from spectrochempy.api.plugins import check_plugin_compatibility
 from spectrochempy.testing.plugins import PluginTestHarness
 
-_HERE = Path(__file__).parent
-
-# Skip all tests if cantera is not installed
 try:
     import cantera as ct
 
     HAS_CANTERA = True
-    # Check if a default mechanism is available
     ct.Solution("gri30.yaml")
     HAS_GRI30 = True
 except Exception:
@@ -39,13 +30,13 @@ except Exception:
 
 
 def test_plugin_metadata():
-    """Plugin has required metadata fields."""
+    """Plugin has required metadata fields (PFR scope only)."""
     plugin = CanteraPlugin()
     assert plugin.name == "cantera"
     assert plugin.version == "0.1.0"
     assert plugin.description
     assert PluginCapability.SIMULATION in plugin.capabilities
-    assert PluginCapability.ANALYSIS in plugin.capabilities
+    assert PluginCapability.ANALYSIS not in plugin.capabilities
     assert PluginCapability.ACCESSOR not in plugin.capabilities
 
 
@@ -53,7 +44,6 @@ def test_plugin_compatibility():
     """Plugin passes full compatibility check."""
     plugin = CanteraPlugin()
     issues = check_plugin_compatibility(plugin)
-    # should report cantera as missing if not installed
     if not HAS_CANTERA:
         assert any("cantera" in issue for issue in issues)
     else:
@@ -66,7 +56,7 @@ def test_plugin_compatibility():
 
 
 def test_registration_with_cantera():
-    """Plugin registers simulations and analyses when cantera is available."""
+    """Plugin registers PFR simulation when cantera is available."""
     if not HAS_CANTERA:
         pytest.skip("cantera not installed")
 
@@ -74,19 +64,14 @@ def test_registration_with_cantera():
     harness.register(CanteraPlugin())
 
     simulations = harness.registry.extensions.list_category("simulation")
-    assert "equilibrium" in simulations
-    assert "reactor_profile" in simulations
-    assert "flame_speed" in simulations
+    assert "PFR" in simulations
 
     analyses = harness.registry.extensions.list_category("analysis")
-    assert "thermo_properties" in analyses
-    assert "kinetic_sensitivity" in analyses
-    assert "spectral_coupling" in analyses
+    assert len(analyses) == 0
 
 
 def test_registration_without_cantera():
     """Plugin is marked FAILED when cantera is missing."""
-    # Simulate by checking the requires check
     harness = PluginTestHarness()
     harness.register(CanteraPlugin())
 
@@ -106,11 +91,16 @@ def test_capability_query():
 
     results = harness.registry.get_by_capability(PluginCapability.SIMULATION)
     names = [r["name"] for r in results]
-    assert "equilibrium" in names
+    assert "PFR" in names
 
 
-def test_package_namespace_exposes_simulations():
-    """Cantera simulations are package-level plugin APIs."""
+# ------------------------------------------------------------------
+# PFR namespace access
+# ------------------------------------------------------------------
+
+
+def test_package_namespace_exposes_pfr():
+    """scp.cantera.PFR is accessible when plugin is registered."""
     if not HAS_CANTERA:
         pytest.skip("cantera not installed")
 
@@ -119,12 +109,11 @@ def test_package_namespace_exposes_simulations():
     if not scp.plugin_manager.has_plugin("cantera"):
         scp.plugin_manager.register(CanteraPlugin())
 
-    assert scp.cantera.equilibrium is equilibrium_composition
-    assert scp.cantera.reactor_profile is reactor_profile
+    assert scp.cantera.PFR is PFR
 
 
 def test_cantera_does_not_register_dataset_accessor():
-    """Cantera has no dataset accessor until dataset semantics are explicit."""
+    """Cantera has no dataset accessor."""
     import spectrochempy as scp
 
     if HAS_CANTERA and not scp.plugin_manager.has_plugin("cantera"):
@@ -132,82 +121,10 @@ def test_cantera_does_not_register_dataset_accessor():
 
     dataset = scp.NDDataset([300.0, 310.0])
     assert not hasattr(dataset, "cantera")
-    assert not hasattr(dataset, "cantera_equilibrium")
 
 
 # ------------------------------------------------------------------
-# Simulation functions
-# ------------------------------------------------------------------
-
-
-@pytest.mark.skipif(not HAS_GRI30, reason="gri30.yaml not available via cantera")
-def test_equilibrium_composition():
-    """Equilibrium composition returns expected structure."""
-    result = equilibrium_composition(
-        mechanism="gri30.yaml",
-        temperature=1200.0,
-        pressure=101325.0,
-        reactants={"CH4": 1.0, "O2": 2.0},
-    )
-
-    assert "species" in result
-    assert "mole_fractions" in result
-    assert "temperature" in result
-    assert result["temperature"] == 1200.0
-    assert result["n_species"] > 0
-    assert result["mole_fractions"].data.ndim >= 1
-
-
-@pytest.mark.skipif(not HAS_GRI30, reason="gri30.yaml not available via cantera")
-def test_thermo_properties():
-    """Thermodynamic properties return expected structure."""
-    result = thermo_properties(
-        mechanism="gri30.yaml",
-        species="CH4",
-        T_min=300.0,
-        T_max=1500.0,
-        n_points=10,
-    )
-
-    assert "temperature" in result
-    assert "cp" in result
-    assert "enthalpy" in result
-    assert "entropy" in result
-    assert "gibbs" in result
-    assert len(result["temperature"].data) == 10
-
-
-@pytest.mark.skipif(not HAS_GRI30, reason="gri30.yaml not available via cantera")
-def test_reactor_profile():
-    """Reactor profile returns expected structure."""
-    result = reactor_profile(
-        mechanism="gri30.yaml",
-        initial_conditions={"T": 1500.0, "P": 101325.0},
-        residence_time=0.01,
-        n_points=10,
-    )
-
-    assert "time" in result
-    assert "temperature" in result
-    assert "species" in result
-    assert "mole_fractions" in result
-    assert len(result["time"]) == 10
-
-
-@pytest.mark.skipif(not HAS_GRI30, reason="gri30.yaml not available via cantera")
-def test_equilibrium_different_reactants():
-    """Equilibrium with different reactant compositions."""
-    result = equilibrium_composition(
-        mechanism="gri30.yaml",
-        temperature=1000.0,
-        pressure=101325.0,
-        reactants={"H2": 2.0, "O2": 1.0},
-    )
-    assert result["n_species"] > 0
-
-
-# ------------------------------------------------------------------
-# PFR tests
+# PFR import and basic checks
 # ------------------------------------------------------------------
 
 
@@ -219,7 +136,7 @@ def test_cantera_is_not_available(monkeypatch):
     assert not _cantera_is_not_available()
 
     with monkeypatch.context() as m:
-        import spectrochempy_cantera._pfr as _pfr  # noqa: PLC0415, PLC0415
+        import spectrochempy_cantera._pfr as _pfr  # noqa: PLC0415
 
         m.setattr(_pfr, "ct", None)
         assert _cantera_is_not_available()
