+++ src/spectrochempy/utils/optional.py
@@ -99,7 +99,7 @@ def import_optional_dependency(
         return None
 
     # Handle submodules: if we have submodule, grab parent module from sys.modules
-    parent = name.split(".")[0]
+    parent = name.split(".", maxsplit=1)[0]
     if parent != name:
         install_name = parent
         module_to_get = sys.modules[install_name]
